% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmvnprd.R
\name{fn}
\alias{fn}
\title{Vectorized computation of product of normal densities for use in pmvnprd}
\usage{
fn(y, z, rho)
}
\description{
Vectorized computation of product of normal densities for use in pmvnprd
}
