% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_i.R
\name{l_i}
\alias{l_i}
\title{Computes the log-likelihood contribution for a subset
of data with fixed x}
\usage{
l_i(data, alpha, beta, sigma2b, sigma2e, q_rule)
}
\arguments{
\item{data}{Dataframe with columns y, x and cens.}

\item{alpha}{Intercept}

\item{beta}{Slope}

\item{sigma2b}{Random effects variance}

\item{sigma2e}{Residual variance}

\item{q_rule}{Quadrature rule}
}
\description{
Computes the log-likelihood contribution for a subset
of data with fixed x
}
