% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_dmnorm.R
\name{log_dmnorm}
\alias{log_dmnorm}
\title{Non-censored likelihood parts}
\usage{
log_dmnorm(y, mu, sigma2e, sigma2b)
}
\arguments{
\item{y}{obs}

\item{mu}{mean}

\item{sigma2e}{residual variance}

\item{sigma2b}{random effect variance}
}
\description{
Non-censored likelihood parts
}
