% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcens.R
\name{mixcens}
\alias{mixcens}
\title{Fit censored (from the left) Linear Mixed-Effects Model}
\usage{
mixcens(x, y, cens, family = "lognormal", use_lmec = FALSE)
}
\arguments{
\item{x}{vector of covariates}

\item{y}{vector of responses}

\item{cens}{vector of censoring indicators, if TRUE y is censored from the left}

\item{family}{lognormal or normal}

\item{use_lmec}{Should lmec::lmec be used for fitting?}
}
\description{
Fit censored (from the left) Linear Mixed-Effects Model
}
\examples{
fit <- with(lindane_landsort, mixcens(year, value, cens))
summary(fit)
plot(fit)

}
