% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmvnprd.R
\name{pmvnprd}
\alias{pmvnprd}
\title{Computes marginalised likelihood contribution}
\usage{
pmvnprd(x, mu, c, v, q_rule)
}
\arguments{
\item{x}{value}

\item{mu}{mean}

\item{c}{covariance}

\item{v}{variance}

\item{q_rule}{Quadrature rule}
}
\description{
Computes marginalised likelihood contribution
}
