% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.mixcens}
\alias{predict.mixcens}
\title{Generic}
\usage{
\method{predict}{mixcens}(object, type = "response", newdata = object$data$x, ...)
}
\arguments{
\item{object}{A mixcens object}
}
\description{
Generic
}
